requirements.txt
# Core Django
Django==5.1.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0
# Production server & static files
gunicorn==22.0.0
whitenoise==6.7.0
# Utilities
python-decouple==3.8
Pillow==11.0.0 #for profile pic